+++ logic-server/src/main/java/logic/server/service/impl/LoginServiceImpl.java
@@ -327,10 +327,10 @@ public class LoginServiceImpl implements ILoginService {
         }
 
         /** 宠物离线时间：单位秒 **/
-        loginResPb.setOfflineTime( (userDTO.getLatestLoginTime().getTime() - userDTO.getLatestLogoutTime().getTime())/1000 );
+        loginResPb.setOfflineTime( (int)((userDTO.getLatestLoginTime().getTime() - userDTO.getLatestLogoutTime().getTime())/1000) );
         /** 宠物离线收益 **/
         long petOfflineIncome = settlementExecutor.petOfflineIncome(userDTO,userAttributeDTO,1);
-        loginResPb.setPetOfflineIncome(petOfflineIncome);
+        loginResPb.setPetOfflineIncome((int)petOfflineIncome);
 
         return loginResPb;
     }
